% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_data.R
\docType{data}
\name{co2_data}
\alias{co2_data}
\title{Atmospheric CO2 (ppm) and atmospheric δ13CO2 data for iWUE calculations}
\format{

}
\source{
https://www.sciencedirect.com/science/article/abs/pii/S1125786520300874
}
\usage{
data(co2_data)
}
\description{
Compiled records of atmospheric CO2 concentrations and stable carbon isotopes to reconstruct climate and derive plant ecophysiological indices from tree rings. Data are from Belmecheri, Lavergne, 2020, Dendrochronologia.
}
\examples{
data(co2_data)
co2_data
plot(co2_data$yr, co2_data$Ca)

}
\references{
Belmecheri, S. & Lavergne, A. (2020). Compiled records of atmospheric
CO2 concentrations and stable carbon isotopes to reconstruct climate and derive
plant ecophysiological indices from tree rings. Dendrochronologia, 63, 125748.
}
\keyword{datasets}
