% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.R
\docType{package}
\name{isocalcR-package}
\alias{isocalcR}
\alias{isocalcR-package}
\title{isocalcR: Isotope Calculations in R}
\description{
Perform common calculations based on published stable isotope theory, such as calculating carbon isotope discrimination and intrinsic water use efficiency from wood or leaf carbon isotope composition. See Farquhar, O'Leary, and Berry (1982) <doi:10.1071/PP9820121>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/justinmathias/isocalcR}
  \item Report bugs at \url{https://github.com/justinmathias/isocalcR/issues}
}

}
\keyword{internal}
