% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piru13C.R
\docType{data}
\name{piru13C}
\alias{piru13C}
\title{piru13C}
\format{
A data frame with 223 rows and 6 variables:
\describe{
  \item{Year}{Year of sample}
  \item{Site}{Study location name}
  \item{wood.d13C}{Measured tree ring (i.e. wood) d13C, in per mille, ‰}
  \item{MGT_C}{Mean growing season temperature, °C}
  \item{Elevation_m}{Elevation of study location, meters}
  \item{frac}{Leaf-to-wood fractionation factor}
}
}
\source{
https://onlinelibrary.wiley.com/doi/full/10.1111/gcb.14273
}
\usage{
data(piru13C)
}
\description{
Accompanying tree ring carbon isotope signature data for "Mathias, J.M. & Thomas, R.B. Disentangling the effects of acidic air pollution, atmospheric CO2, and climate change
on recent growth of red spruce trees in the Central Appalachian Mountains. Glob. Chang. Biol. 24, 3938–3953 (2018).".
}
\examples{


data(piru13C)
head(piru13C)
}
\references{
Mathias, J.M. & Thomas, R.B. Disentangling the effects of acidic air pollution, atmospheric CO2, and climate change
on recent growth of red spruce trees in the Central Appalachian Mountains. Glob. Chang. Biol. 24, 3938–3953 (2018).
}
\keyword{datasets}
